Semi-Interactive Team Management

Add a team using a Slack response code.

add_team_code(code, redirect_uri = NULL, verbose = TRUE)

Arguments

code
character, a code returned by the slack oauth2 v2 api.

redirect_uri
character, the uri to which the user was redirected when the code was generated.

verbose
logical, Print messages to console, Default: TRUE

Value

The token (invisibly)

Details

Launch a browser window to interactively grant slackteams permission to act on your behalf on a Slack team.

Note

This function does not currently work in an Rstudio Server setup. We are exploring options to remedy this situation.