Interactive Team Managment

Add a team interactively

add_team_interactive(scopes = load_scopes(), verbose = TRUE)

Arguments

scopes
character, scopes to request. Must include "users:read", "channels:read", "groups:read", "im:read", and "mpim:read" at minimum.

verbose
logical, Print messages to console, Default: TRUE

Details

Launch a browser window to interactively grant slackteams permission to act on your behalf on a Slack team.

Note

This function does not currently work in an Rstudio Server setup. We are exploring options to remedy this situation.