Custom App Helper

Create a manifest for a custom Slack App. See the Slack Apps vignette for details about using custom Slack Apps.

create_custom_app(
  app_name,
  redirect_urls = NULL,
  description = NULL,
  long_description = NULL
)

Arguments

app_name

character; a name to give to your custom app.

redirect_urls

character; urls to which the app can be redirected during authentication. We recommend including localhost on a specific port (eg, http://127.0.0.1:4242) for testing, and often a shinyapps.io url.

description

character; an optional one-sentence description of your app. If this parameter is NULL, an adequate description is provided.

long_description

character; an optional paragraph (or more) giving more details about your app. This field supports markdown. If this parameter is NULL, an adequate long description is provided.

Value

The app manifest as an editable R object, invisibly.

Examples

create_custom_app(
  app_name = "my_cool_app",
  redirect_urls = c("http://127.0.0.1:4242", "myapp.shinyapps.io")
)
#> Visit https://api.slack.com/apps & Create New App
#> Copy/paste this manifest:
#> Warning: Slack may flag your manifest with the error 'Must provide an array.'
#> If that is still occurring, move the value after : to a new line,
#> indented to the level of the setting, and then add '-'.
#> _metadata:
#>   major_version: 1
#>   minor_version: 0
#> display_information:
#>   name: my_cool_app
#>   long_description: This app is used by the [{slackteams} R package](https://github.com/yonicd/slackteams)
#>     to connect R to Slack workspaces. Use the app to authenticate yourself and generate
#>     a token, which allows you to interact with the Slack API. See the {slackteams}
#>     documentation for more information.
#>   description: A custom app for use with the {slackteams} R package.
#>   background_color: '#000000'
#> settings:
#>   org_deploy_enabled: false
#>   socket_mode_enabled: false
#>   token_rotation_enabled: false
#> oauth_config:
#>   redirect_urls:
#>     - http://127.0.0.1:4242
#>     - myapp.shinyapps.io
#>   scopes:
#>     user:
#>       - team:read
#>       - channels:read