Prepare an App Manifest

Convert an app manifest list object to a yaml for upload.

custom_app_yaml(app_manifest)

Arguments

app_manifest

list; an app manifest such as the one returned by create_custom_app.

Value

The manifest in yaml format (invisibly).

Examples

my_manifest <- create_custom_app("myApp")
#> Visit https://api.slack.com/apps & Create New App
#> Copy/paste this manifest:
#> _metadata:
#>   major_version: 1
#>   minor_version: 0
#> display_information:
#>   name: myApp
#>   long_description: This app is used by the [{slackteams} R package](https://github.com/yonicd/slackteams)
#>     to connect R to Slack workspaces. Use the app to authenticate yourself and generate
#>     a token, which allows you to interact with the Slack API. See the {slackteams}
#>     documentation for more information.
#>   description: A custom app for use with the {slackteams} R package.
#>   background_color: '#000000'
#> settings:
#>   org_deploy_enabled: no
#>   socket_mode_enabled: no
#>   token_rotation_enabled: no
#> oauth_config:
#>   redirect_urls: http://127.0.0.1:4242
#>   scopes:
#>     user: team:read
my_manifest$`_metadata`$minor_version <- 1L
custom_app_yaml(my_manifest)
#> _metadata:
#>   major_version: 1
#>   minor_version: 1
#> display_information:
#>   name: myApp
#>   long_description: This app is used by the [{slackteams} R package](https://github.com/yonicd/slackteams)
#>     to connect R to Slack workspaces. Use the app to authenticate yourself and generate
#>     a token, which allows you to interact with the Slack API. See the {slackteams}
#>     documentation for more information.
#>   description: A custom app for use with the {slackteams} R package.
#>   background_color: '#000000'
#> settings:
#>   org_deploy_enabled: no
#>   socket_mode_enabled: no
#>   token_rotation_enabled: no
#> oauth_config:
#>   redirect_urls: http://127.0.0.1:4242
#>   scopes:
#>     user: team:read