Get Channel List

Fetch channel information from the Slack API.

get_conversations_list(
  type = c("public_channel", "private_channel", "mpim", "im"),
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf
)

Arguments

type

character, the types of channels to include in the fetch.

...

additional parameters passed on to post_slack

token

character, api token issued by slack

max_results

integer, the maximum number of results to return (total).

max_calls

integer, the maximum number of separate API calls to make while constructing the response.

Value

tibble